% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sentiment_app.R
\name{run_app_sentiment_dict}
\alias{run_app_sentiment_dict}
\title{run_app_sentiment_dict}
\usage{
run_app_sentiment_dict(.cache_pointer = NULL)
}
\arguments{
\item{.cache_pointer}{a \code{cachem::cache_*} object}
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
1 + 1
}
